:recycle: Move audio and video nodes to own plugins
