+++ src/plugins/paste-url-handler/handlers/AudioHandler.js
@@ -1,5 +1,3 @@
-import React from "react";
-
 import mimeLookup from "browser-media-mime-type";
 
 export default function AudioHandler(options) {
@@ -10,20 +8,10 @@ export default function AudioHandler(options) {
       insertAudio,
     },
     helpers: {},
-    components: {
-      AudioNode,
-    },
-    plugins: [RenderAudioNode, { schema }],
+    components: {},
+    plugins: [],
   };
 }
-
-const schema = {
-  blocks: {
-    audio: {
-      isVoid: true,
-    },
-  },
-};
 
 const validate = url => {
   const fileExtension = url.substr(url.lastIndexOf(".") + 1);
@@ -44,45 +32,3 @@ function insertAudio(change, src, mimeType, target) {
     data: { src, mimeType },
   });
 }
-
-function AudioNode(props) {
-  const { src, mimeType, selected, ...attributes } = props;
-
-  return (
-    <div
-      className={`plugin-wrapper ${selected ? "selected" : ""}`}
-      {...attributes}
-    >
-      <audio
-        className="audio"
-        controls
-        preload="metadata"
-        style={{ width: "100%", minHeight: "54px" }}
-      >
-        <source src={src} type={mimeType} />
-        Your browser does not support the audio tag.
-      </audio>
-    </div>
-  );
-}
-
-const RenderAudioNode = {
-  renderNode(props) {
-    const { attributes, node, isFocused } = props;
-
-    if (node.type === "audio") {
-      const src = node.data.get("src");
-      const mimeType = node.data.get("mimeType");
-      if ((mimeType || "").includes("audio")) {
-        return (
-          <AudioNode
-            src={src}
-            mimeType={mimeType}
-            selected={isFocused}
-            {...attributes}
-          />
-        );
-      }
-    }
-  },
-};
