+++ src/plugins/paste-url-handler/handlers/VideoHandler.js
@@ -1,8 +1,6 @@
-import React from "react";
-
 import mimeLookup from "browser-media-mime-type";
 
-export default function YoutubeHandler(options) {
+export default function VideoHandler(options) {
   return {
     validate,
     dealWithIt,
@@ -10,20 +8,10 @@ export default function YoutubeHandler(options) {
       insertVideo,
     },
     helpers: {},
-    components: {
-      VideoNode,
-    },
-    plugins: [RenderVideoNode, { schema }],
+    components: {},
+    plugins: [],
   };
 }
-
-const schema = {
-  blocks: {
-    video: {
-      isVoid: true,
-    },
-  },
-};
 
 const validate = url => {
   const fileExtension = url.substr(url.lastIndexOf(".") + 1);
@@ -44,40 +32,3 @@ function insertVideo(change, src, mimeType) {
     data: { src, mimeType },
   });
 }
-
-function VideoNode(props) {
-  const { src, mimeType, selected, ...attributes } = props;
-
-  return (
-    <div
-      className={`plugin-wrapper ${selected ? "selected" : ""}`}
-      {...attributes}
-    >
-      <video className="video" controls preload="metadata">
-        <source src={src} type={mimeType} />
-        Your browser does not support the video tag.
-      </video>
-    </div>
-  );
-}
-
-const RenderVideoNode = {
-  renderNode(props) {
-    const { attributes, node, isFocused } = props;
-
-    if (node.type === "video") {
-      const src = node.data.get("src");
-      const mimeType = node.data.get("mimeType");
-      if ((mimeType || "").includes("video")) {
-        return (
-          <VideoNode
-            src={src}
-            mimeType={mimeType}
-            selected={isFocused}
-            {...attributes}
-          />
-        );
-      }
-    }
-  },
-};
